refactor: move TypeIconComponent from features to core
