+++ issue-body.md
@@ -1,72 +1,21 @@
 ## 概要
-新しい難易度「コイキング級」を追加する。このモードでは、単タイプのポケモンのみが出題され、初心者でも遊びやすい難易度となる。
+dependency-cruiserがレイヤー境界違反を検出。`core/`が`features/`に依存している箇所を修正する。
 
-## 背景
-現在の難易度は以下の3つ：
-- **EASY**: 倍率表示あり、相性表閲覧可能
-- **NORMAL**: 標準ルール、10ターン
-- **MASTER**: ハードコア、6ターン、Solve1回のみ
+## 違反内容
+```
+core → features 違反:
+src/app/core/components/chip/chip.component.ts
+    → src/app/features/type-icon/type-icon.component.ts
+```
 
-初心者にとって、複合タイプの相性計算は複雑で難しい。単タイプのみに限定することで、より学習しやすく、ゲームに慣れるための入門モードが必要。
+## 対応方針
+`TypeIconComponent`を`features/`から`core/components/`に移動する。
 
-## 新しい難易度体系
-難易度をポケモン名で表現し、4段階に拡張する：
+### 理由
+- `TypeIconComponent`は`core/`から使われている
+- 再利用可能な汎用コンポーネントは`core/`に属すべき
 
-1. **コイキング級** 🐟 - 単タイプのみ（新規追加）
-2. **ピチュー級** ⚡ - EASY相当（既存のEASYをリネーム）
-3. **ウインディ級** 🔥 - NORMAL相当（既存のNORMALをリネーム）
-4. **ガブリアス級** 🐉 - MASTER相当（既存のMASTERをリネーム）
-
-## コイキング級の仕様
-
-### 出題ルール
-- **対象**: 単タイプのポケモンのみ（18タイプ）
-- **複合タイプ**: 出題されない
-- **ターン数**: 10ターン（NORMAL相当）
-- **Solveモード**: 制限なし（何度でも使用可能）
-
-### UI表示
-- **倍率表示**: なし（単タイプなので×4と×0.25が存在せず、⭕/⚪/🔺/✖で十分判断可能）
-- **相性表**: いつでも閲覧可能（EASY相当）
-- **難易度表示**: 「コイキング級」+ ドット絵アイコン
-- **判定アイコン**: ⭕（×2）、⚪（×1）、🔺（×0.5）、✖（×0）
-
-### ゲームロジック
-- 既存の`generateTarget()`関数を修正し、難易度が`magikarp`の場合は単タイプのみを生成
-- 複合タイプの組み合わせを除外
-
-## 実装タスク
-
-### フロントエンド
-- [ ] 難易度の型定義を更新（`'easy' | 'normal' | 'master'` → `'magikarp' | 'pichu' | 'arcanine' | 'garchomp'`）
-- [ ] Settingsモーダルに「コイキング級」を追加
-- [ ] 難易度表示を新しい名前に更新（全画面）
-- [ ] ドット絵アイコンの準備と表示（Issue #19と連携）
-
-### ゲームロジック
-- [ ] `generateTarget()`関数を修正し、`magikarp`モードで単タイプのみ生成
-- [ ] 難易度ごとのルール設定を更新
-  - `magikarp`: 10ターン、倍率表示なし、相性表あり、単タイプのみ
-  - `pichu`: 10ターン、倍率表示あり、相性表あり
-  - `arcanine`: 10ターン、倍率表示なし
-  - `garchomp`: 6ターン、倍率表示なし、Solve1回のみ
-
-### テスト
-- [ ] 単タイプのみが生成されることを確認
-- [ ] 各難易度のルールが正しく適用されることを確認
-- [ ] 難易度の切り替えが正常に動作することを確認
-
-## 完了条件
-- [ ] コイキング級が選択可能
-- [ ] コイキング級では単タイプのポケモンのみが出題される
-- [ ] 判定結果が⭕/⚪/🔺/✖のアイコンで表示される（倍率表示なし）
-- [ ] 相性表が利用可能
-- [ ] 難易度表示が新しいポケモン名で表示される
-- [ ] すべての難易度で正常にゲームがプレイできる
-
-## 関連Issue
-- #19: 難易度表示のUI/UX改善（ドット絵アイコンの実装）
-
-## 備考
-- ドット絵アイコンは著作権に配慮し、フリー素材または自作のものを使用する
-- 将来的に難易度を追加する可能性を考慮し、拡張しやすい設計にする
+## タスク
+- [ ] `TypeIconComponent`を`core/components/`に移動
+- [ ] import文を更新
+- [ ] `npm run dep-check`で違反が解消されたことを確認
