+++ src/app/core/components/type-icon/type-icon.component.html
